requirements.txt
asgiref==3.6.0
astroid==2.14.2
bcrypt==4.0.1
boto3==1.26.79
botocore==1.29.79
click==8.1.3
dill==0.3.6
Django==4.1.7
django-cors-headers==3.13.0
django-environ==0.9.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
factory-boy==3.2.1
Faker==17.3.0
gunicorn==20.1.0
isort==5.12.0
jmespath==1.0.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.0.0
psycopg2==2.9.5
Pygments==2.14.0
PyJWT==2.6.0
pylint==2.16.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-dateutil==2.8.2
pytz==2022.7.1
redgreenunittest==0.1.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
urllib3==1.26.14
wrapt==1.14.1
